# wildberriesapi/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is the API documentation for the WB Analytics Collector Service.",
        "title": "WB Analytics Collector Service API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/api/incomes": {
            "get": {
                "description": "Возвращает поставки за период",
                "tags": [
                    "Incomes"
                ],
                "summary": "Получить поставки из WB API",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Дата начала (YYYY-MM-DD)",
                        "name": "dateFrom",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object",
                                "additionalProperties": true
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/orders": {
            "get": {
                "description": "Возвращает список заказов за указанный период",
                "tags": [
                    "Orders"
                ],
                "summary": "Получить заказы из WB API",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Дата начала (YYYY-MM-DD)",
                        "name": "dateFrom",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Дата окончания (YYYY-MM-DD)",
                        "name": "dateTo",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object",
                                "additionalProperties": true
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/paid_storage/download": {
            "get": {
                "description": "Метод возвращает отчёт о платном хранении по ID задания на генерацию.",
                "tags": [
                    "Paid Storage"
                ],
                "summary": "Получить отчёт из WB API",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Дата начала (YYYY-MM-DD)",
                        "name": "dateFrom",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Дата окончания (YYYY-MM-DD)",
                        "name": "dateTo",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object",
                                "additionalProperties": true
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/paid_storage/start": {
            "get": {
                "description": "Метод создаёт задание на генерацию отчёта о платном хранении.",
                "tags": [
                    "Paid Storage"
                ],
                "summary": "Создать отчёт из WB API",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Дата начала (YYYY-MM-DD)",
                        "name": "dateFrom",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Дата окончания (YYYY-MM-DD)",
                        "name": "dateTo",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object",
                                "additionalProperties": true
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/paid_storage/storage": {
            "get": {
                "description": "Возвращает статус задания на генерацию отчёта о платном хранении заказов за указанный период",
                "tags": [
                    "Paid Storage"
                ],
                "summary": "Проверить статус из WB API",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object",
                                "additionalProperties": true
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/sales": {
            "get": {
                "description": "Возвращает продажи за период",
                "tags": [
                    "Sales"
                ],
                "summary": "Получить продажи из WB API",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Дата начала (YYYY-MM-DD)",
                        "name": "dateFrom",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object",
                                "additionalProperties": true
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/stocks": {
            "get": {
                "description": "Возвращает текущие остатки",
                "tags": [
                    "Stocks"
                ],
                "summary": "Получить остатки из WB API",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Дата начала (YYYY-MM-DD)",
                        "name": "dateFrom",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object",
                                "additionalProperties": true
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/tariffs": {
            "get": {
                "description": "Метод возвращает данные о комиссии WB по родительским категориям товаров согласно модели продаж.",
                "tags": [
                    "Tariffs"
                ],
                "summary": "Получить Комиссия по категориям товаров из WB API",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object",
                                "additionalProperties": true
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/tariffs/box": {
            "get": {
                "description": "Метод возвращает данные о комиссии WB по родительским категориям товаров согласно модели продаж.",
                "tags": [
                    "Tariffs"
                ],
                "summary": "Получить Комиссия по категориям товаров из WB API",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Дата (YYYY-MM-DD)",
                        "name": "dateFrom",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object",
                                "additionalProperties": true
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/tariffs/pallet": {
            "get": {
                "description": "Метод возвращает данные о комиссии WB по родительским категориям товаров согласно модели продаж.",
                "tags": [
                    "Tariffs"
                ],
                "summary": "Получить Комиссия по категориям товаров из WB API",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Дата (YYYY-MM-DD)",
                        "name": "dateFrom",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object",
                                "additionalProperties": true
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    }
}